Request sensor states when shuttles start
Request sensor states when shuttles start.  This fixes the problem of the train setting off in the wrong direction when at one end, when it doesn't know the state of the end sensor at shuttle start. Also set direction sensors in use as fixed forward direction uses forward sensor, as the attempt to auto determine this could also cause the train to set off in the wrong direction at startup.